+++ FERRETERIA  Joel/Models/Categoria.cs
@@ -2,7 +2,7 @@ namespace FERRETERIA__Joel.Models
 {
     public class Categoria
     {
-        public short IdCategoria { get; set; }
+        public int IdCategoria { get; set; }
         public string Codigo { get; set; } = string.Empty;
         public string Nombre { get; set; } = string.Empty;
         public string? Descripcion { get; set; }
@@ -10,6 +10,6 @@ namespace FERRETERIA__Joel.Models
         public byte Estado { get; set; }
         public DateTime FechaRegistro { get; set; }
         public DateTime? FechaActualizacion { get; set; }
-        public short IdEmpleadoResponsable { get; set; }
+        public int IdEmpleadoResponsable { get; set; }
     }
 }
