+++ FERRETERIA  Joel/Models/Empleado.cs
@@ -2,7 +2,7 @@ namespace FERRETERIA__Joel.Models
 {
     public class Empleado
     {
-        public short IdEmpleado { get; set; }
+        public int IdEmpleado { get; set; }
         public string Nombre { get; set; } = string.Empty;
     }
 }
